requirements.txt
pytest-mockito==0.0.4
pandas==1.4.4
xmltodict==0.13.0
requests==2.22.0
pytest==7.1.3